Revert "fix: PR review #3"
This reverts commit 230013c1ba51271b6898c39c1b28ed345061932f.
