+++ src/components/Action/index.tsx
@@ -1,9 +1,8 @@
 import { Button, ButtonProps } from "./Button";
 import { Link, LinkProps } from "./Link";
-import { LinkWrapper, LinkWrapperProps } from "./LinkWrapper";
 
 interface ActionProps {
-  children: React.ReactElement<ButtonProps | LinkProps | LinkWrapperProps>[];
+  children: React.ReactElement<ButtonProps | LinkProps>[];
 }
 
 export const Action = ({ children }: ActionProps) => {
@@ -12,4 +11,3 @@ export const Action = ({ children }: ActionProps) => {
 
 Action.Button = Button;
 Action.Link = Link;
-Action.LinkWrapper = LinkWrapper;
