+++ src/components/Actions/index.tsx
@@ -1,9 +1,8 @@
 import { NonEmptyArr } from "../../types/utils";
 import { LinkProps } from "../Action/Link";
-import { LinkWrapperProps } from "../Action/LinkWrapper";
 import { ButtonProps } from "../Button";
 
-type ActionsElement = React.ReactElement<LinkProps | ButtonProps | LinkWrapperProps>;
+type ActionsElement = React.ReactElement<LinkProps | ButtonProps>;
 type ActionsChildren = NonEmptyArr<ActionsElement>;
 
 export interface ActionsProps {
