+++ src/components/List/Item.tsx
@@ -1,5 +1,4 @@
-import { Action } from "../Action";
-import { Actions, ActionsProps } from "../Actions";
+import { ActionsProps } from "../Actions";
 import { Icon, IconEnum } from "../Icon";
 
 type ListType = "card" | "text" | "bullet" | "cover";
@@ -44,27 +43,25 @@ export const ListItem = ({ type, title, subtitle, additional, count, image, inde
     );
   }
 
-  if (type === "text" && actions) {
-
+  if (type === "text") {
     return (
-      // <actions>
-        <div className="flex flex-col items-start gap-3 w-full pb-6">
-          <div className="flex items-start gap-3">
-            <div className="flex h-5 w-0.5 rounded-[1px] bg-zinc-200"></div>
-            <p className="flex-grow text-sm text-left text-zinc-400">{additional}</p>
-          </div>
-          <div className="flex items-center gap-2 w-full">
-            <p className="flex-grow text-base font-semibold text-left text-zinc-900 dark:text-zinc-100">{title}</p>
-            <div className="flex items-center gap-1">
-              <div className="flex flex-col justify-center items-center h-6 w-6">
-                {count?.icon && <Icon name={count.icon} color="neutral" size="sm" />}
-              </div>
-              {count && <p className="text-sm font-medium text-left text-zinc-400">{count.value}</p>}
-            </div>
-          </div>
-          <p className="text-sm text-left text-zinc-600 dark:text-zinc-400">{subtitle}</p>
+      <div className="flex flex-col items-start gap-3 w-full pb-6">
+        <div className="flex items-start gap-3">
+          <div className="flex h-5 w-0.5 rounded-[1px] bg-zinc-200"></div>
+          <p className="flex-grow text-sm text-left text-zinc-400">{additional}</p>
         </div>
-      // </actions>
+        <div className="flex items-center gap-2 w-full">
+          <p className="flex-grow text-base font-semibold text-left text-zinc-900 dark:text-zinc-100">{title}</p>
+          <div className="flex items-center gap-1">
+            <div className="flex flex-col justify-center items-center h-6 w-6">
+              {count?.icon && <Icon name={count.icon} color="neutral" size="sm" />}
+            </div>
+            {count && <p className="text-sm font-medium text-left text-zinc-400">{count.value}</p>}
+          </div>
+        </div>
+        <p className="text-sm text-left text-zinc-600 dark:text-zinc-400">{subtitle}</p>
+        {actions && actions}
+      </div>
     );
   }
