+++ src/stories/Page.stories.tsx
@@ -1,12 +1,8 @@
 import { Story, Meta } from "@storybook/react/types-6-0";
 import { Action } from "../components/Action";
 import { Block } from "../components/Block";
-import { List } from "../components/List";
-import { Card } from "../components/Card";
 import { Page, PageProps } from "../components/Page";
-import { Pills } from "../components/Pills";
 import { blogPosts, experience, openSource, personal, projects, spotify, university } from "./_data";
-import { Seperator } from "../components/Seperator";
 
 export default {
   title: "Private/Page",
@@ -33,9 +29,9 @@ const argObject = {
   main: [
     <>
       <Block title="My writings" imagePath="https://avatars.githubusercontent.com/u/65030610?s=200&v=4" size={2} isEditable={true} actions={<Action.Link url="https://dev.to"/>}>
-        <List>
+        <Block.List>
           {blogPosts.map((post, index) => (
-            <List.Item
+            <Block.List.Item
               type={post.type}
               title={post.title}
               additional={post.additional}
@@ -43,13 +39,13 @@ const argObject = {
               key={`Block-${index}`}
             />
           ))}
-        </List>
+        </Block.List>
       </Block>
       <Block title="Open Source" imagePath="https://avatars.githubusercontent.com/u/9919?s=200&v=4" size={2} isEditable={true} actions={<Action.Link url="https://github.com"/>}>
-        <Pills pills={["react.js", "vue.js", "angular.js", "ember.js", "svelte.js"]} />
-        <List>
+        <Block.Pills pills={["react.js", "vue.js", "angular.js", "ember.js", "svelte.js"]} />
+        <Block.List>
           {openSource.map((project, index) => (
-            <List.Item
+            <Block.List.Item
               type={project.type}
               title={project.title}
               additional={project.additional}
@@ -58,12 +54,12 @@ const argObject = {
               key={`Block-${index}`}
             />
           ))}
-        </List>
+        </Block.List>
       </Block>
       <Block imagePath="https://avatars.githubusercontent.com/u/24602613?s=200&v=4" size={2} isEditable={true} title="Deployed Projects">
-        <List>
+        <Block.List>
           {projects.map((project, index) => (
-            <List.Item
+            <Block.List.Item
               type={project.type}
               title={project.title}
               additional={project.additional}
@@ -72,19 +68,19 @@ const argObject = {
               key={`Block-${index}`}
             />
           ))}
-        </List>
+        </Block.List>
       </Block>
       <Block title="My Music" imagePath="https://avatars.githubusercontent.com/u/251374?s=200&v=4" size={2} isEditable={true} actions={<Action.Link url="https://spotify.com"/>}>
-        <Card
+        <Block.Card
           type="horizontal"
           title="This album title"
           subtitle="Artist Name"
           image="https://placeimg.com/640/480/arch"
           icon="PlayIcon"
         />
-        <List>
+        <Block.List>
           {spotify.map((album, index) => (
-            <List.Item
+            <Block.List.Item
               index={index + 1}
               type={album.type}
               title={album.title}
@@ -94,16 +90,16 @@ const argObject = {
               key={`Block-${index}`}
             />
           ))}
-        </List>
+        </Block.List>
       </Block>
     </>,
   ],
   aside: [
     <>
       <Block title="Experience" imagePath="https://avatars.githubusercontent.com/u/357098?s=200&v=4" size={1} isEditable={true} actions={<Action.Link url="https://linkedin.com"/>}>
-        <List>
+        <Block.List>
           {experience.map((job, index) => (
-            <List.Item
+            <Block.List.Item
               type={job.type}
               title={job.subtitle}
               subtitle={job.title}
@@ -112,11 +108,11 @@ const argObject = {
               key={`Block-${index}`}
             />
           ))}
-        </List>
-        <Seperator />
-        <List>
+        </Block.List>
+        <Block.Seperator />
+        <Block.List>
           {university.map((job, index) => (
-            <List.Item
+            <Block.List.Item
               type={job.type}
               title={job.title}
               subtitle={job.subtitle}
@@ -125,7 +121,7 @@ const argObject = {
               key={`Block-${index}`}
             />
           ))}
-        </List>
+        </Block.List>
       </Block>
     </>,
   ],
